weather_project/requirements.txt
fastapi
uvicorn
jinja2
tensorflow
numpy
Pillow
python-multipart
